Add method for Area to know the points of interest it contains.

+++ src/main/scala/com/github/shafiquejamal/point/AreaPopulation.scala
@@ -1,3 +1,3 @@
 package com.github.shafiquejamal.point
 
-case class AreaPopulation[T](area: Area[T], population: Int)
+case class AreaPopulation[T](area: Area[T], population: Long)
